fixes #2
Add application icon and refine project file formatting

A new `<PropertyGroup>` hath been added to the project file,
bestowing an `<ApplicationIcon>` that doth specify the path
to the icon for our noble application. Furthermore, the
indentation of the `<PackageReference>` for `Newtonsoft.Json`
hath been adjusted, transforming a leading tab into spaces
for the sake of consistency and harmony in our formatting.