Fixes #8
Refactor template file path handling

The code hath been altered to eschew the use of a hardcoded path for reading the template file. Instead, it now doth construct the path dynamically, utilizing `AppDomain.CurrentDomain.BaseDirectory`. This noble change dost enhance the flexibility and portability of the code, thus avoiding the perils of fixed paths that may lead to misfortune.