Fixes #10
Refine list filtering in GetFirstActivelistWithName

In the `Simple_image_server` namespace, the method
`GetFirstActivelistWithName` hath been altered to enhance
the filtering condition. The previous check for an exact
match of `Name` with `filePath` hath been replaced.
Now, it doth verify if any segment of the `Name`,
when divided by ';', aligns with `filePath`, whilst
still ensuring the list's active status and compliance
with the criteria of active days and time.