fixes #14
Enhance ListBox rendering and item display

This commit bestows upon the `lbElementsInList` the power of custom rendering by setting its drawing mode to `OwnerDrawFixed` and binding it to the `LbElementsInList_DrawItem` event. Within this method, we ensure the validity of the index, retrieve the item, and artfully display its name alongside its index, all while drawing a focus rectangle to guide the eye.

Moreover, we enrich the experience by incorporating logic to reveal the count of images linked to items in `lbLists`, thus elevating the visual representation of our lists.